# server/swagger.openapi3.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/user/current": {
      "get": {
        "operationId": "UserController_getCurrentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCurrentResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ]
      },
      "patch": {
        "operationId": "UserController_updateCurrentUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCurrentUpdateRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCurrentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/api/user/current/avatar": {
      "post": {
        "operationId": "UserController_updateCurrentUserAvatar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "422": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/api/user/avatar/{fileName}": {
      "get": {
        "operationId": "UserController_getUserAvatar",
        "parameters": [
          {
            "name": "fileName",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/api/user/reset-password": {
      "patch": {
        "operationId": "UserController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserResetPasswordRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/api/user/new-password": {
      "patch": {
        "operationId": "UserController_newPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserNewPasswordRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/api/user/change-password": {
      "patch": {
        "operationId": "UserController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserChangePasswordRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ]
      }
    },
...
  },
  "info": {
    "title": "MeshHub Swagger",
    "description": "API приложения MeshHub",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "x-access-token"
      }
    },
    "schemas": {
      "CgSoftResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
      "UserMetaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "aboutYourself": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "favoriteSoft": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CgSoftResponse"
            }
          }
        },
        "required": [
          "id"
        ]
      },
      "UserCurrentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "middleName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "isConfirmed": {
            "type": "boolean"
          },
          "meta": {
            "$ref": "#/components/schemas/UserMetaResponseDto"
          }
        },
        "required": [
          "id",
          "email",
          "isConfirmed",
          "meta"
        ]
      },
      "CgSoftRequest": {
        "type": "object",
        "properties": {
          "id": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ]
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "UserCurrentUpdateRequestDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "middleName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "aboutYourself": {
            "type": "string"
          },
          "favoriteSoft": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CgSoftRequest"
            }
          }
        }
      },
      "UserResetPasswordRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "email"
        ]
      },
      "UserNewPasswordRequestDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "password": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "password",
          "confirmPassword"
        ]
      },
      "UserChangePasswordRequestDto": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          }
        },
        "required": [
          "oldPassword",
          "password",
          "confirmPassword"
        ]
      },
...
    }
  }
}